Close some issue open in codeclimate

+++ src/Service/Cache/CacheItemPool.php
@@ -1,11 +1,21 @@
 <?php
 namespace Service\Cache;
+
 use \Helper\DateHelper;
 
 class CacheItemPool implements CacheItemPoolInterface
 {
     //Quaue saved will be elaborate by commit method.
     protected $queueSaved = array();
+
+    //Save the cacheDir by CacheDir::getCacheDir().
+    private $cacheDir;
+
+    //CacheItemPool constructor!
+    public function __construct()
+    {
+        $this->cacheDir = CacheDir::getCacheDir();
+    }
 
     //Must return an istance of CacheItem.
     public function getItem($key)
@@ -37,12 +47,12 @@ class CacheItemPool implements CacheItemPoolInterface
 
     public function hasItem($key)
     {
-        return file_exists(CacheDir::getCacheDir() . $key);
+        return file_exists($this->cacheDir . $key);
     }
 
     public function clear()
     {
-        if (empty($files = glob(CacheDir::getCacheDir() . '*'))) {
+        if (empty($files = glob($this->cacheDir . '*'))) {
             return false;
         }
 
@@ -57,7 +67,7 @@ class CacheItemPool implements CacheItemPoolInterface
 
     public function deleteItem($key)
     {
-        return $this->hasItem($key) && unlink(CacheDir::getCacheDir() . $key);
+        return $this->hasItem($key) && unlink($this->cacheDir . $key);
     }
 
     public function deleteItems(array $keys)
@@ -79,7 +89,7 @@ class CacheItemPool implements CacheItemPoolInterface
 
         $toWrite = '<?php $item=array(\'value\' => ' . var_export($cacheItem->get(), true) . ', \'expire\' => ' . var_export($cacheItem->getExpires(), true) . '); ?>';
 
-        return ($fileCache = fopen(CacheDir::getCacheDir() . $cacheItem->getKey(), 'w')) &&
+        return ($fileCache = fopen($this->cacheDir . $cacheItem->getKey(), 'w')) &&
                 fwrite($fileCache, $toWrite) &&
                 fclose($fileCache);
     }
@@ -114,21 +124,26 @@ class CacheItemPool implements CacheItemPoolInterface
         return !$cacheItem->isKeyEmpty() && !$cacheItem->isValueEmpty();
     }
 
+    //Inizialize item if it's in cache.
+    //Return true if inizialize in OK, or false if item cache is expire.
     protected function initializeItem(CacheItem $cacheItem)
     {
         $key = $cacheItem->getKey();
-        include CacheDir::getCacheDir() . $key;
+        include $this->cacheDir . $key;
         $itemExpire = $item['expire'];
         $expire = $itemExpire === null ? $itemExpire : date_create()->setTimestamp($itemExpire);
 
         //Check if cache item is valid or not.
-        if ($expire === null || DateHelper::isDateInFuture($expire)) {
-            $cacheItem->set($item['value']);
-            if ($expire !== null) {
-                $cacheItem->expiresAt($expire);
-            }
-        } else {
+        if ($expire !== null && !DateHelper::isDateInFuture($expire)) {
             $this->deleteItem($key);
+            return false;
         }
+
+        $cacheItem->set($item['value']);
+        if ($expire !== null) {
+            $cacheItem->expiresAt($expire);
+        }
+
+        return true;
     }
 }
